Migrate Query.testExternalSchemaComposition
